brunel@ham.esil.univ-mrs.fr
This patch cleans up the printing of beta version notices
(which was used before the release of 1.7) so that they
are now controlled by a define IS_BETA_VERSION, in
configure.in (and also in shared.h, for non-configure builds

This patch replaces all the current capability tags by
the single tag "+1.8pre1".  The idea is that for 1.8.0
the capability string will be "+1.8", and this patch
is to test things properly before release.

I also changed researched and researchpoints into 32 bit
ints, to avoid possible over/under-flows.

Patches written by David Pfitzner <dwp@mso.anu.edu.au>.