Many changes, for partial implementation of governments rulesets, based on patches submitted by Sune Kirkeby <sune@interspace.dk>, but with some substatial changes by me.  This also includes (but currently #ifdef-ed out) code based on patch submitted by Rizos Sakellariou <rizos@cs.rice.edu>, to dynamically evaluate government types.
This change is intended to be a no-op in terms of actual effects
when using the default rulesets; some changes to actual effects
should follow.  The code to dynamically evaluate governments is
included in this patch, but currently #ifdef-ed out for regression
testing, and because it needs updating.  See list in
common/government.c for other things which still need doing.