Add new feature to server: unique citynames.
Activated through the "set citynames X" server option. If set to 0,
there are no restrictions. If set to 1 (new default), city names have
to be unique to a player. If set to 2 or 3, city names have to be
globally unique. If set to 3, a player isn't allowed to use a default
city name of another nations and it must also be globally unique.

Patch by tkurppa <teemu@hoasnet.fi>