Fix savegame and loadgame code so that the dependency on the order of nations in data/default/nations.ruleset is eliminated.  Player nations are now stored by name, as a "nation" attribute, rather than only by numeric nation index saved as a "race" attribute.
Code to interpret the "race" attribute is still present, runs if no
"nation" attribute is found, and indirects through a namelist derived
from the order of nations in the 1.14.0 nations ruleset. Thus this
change does not break reading old savefiles.

Savefiles written by the new savegame() may even be readable by older
versions, because the numeric index is still saved as "race".  The
nation assignments in the resulting save may be scrambled when reread,
though.