AI Diplomacy. You can now do diplomacy with AI players as if they were human players, and AI players will sometimes contact human players and each other with proposals. In team mode, AI players will cooperate tighthly.
The AIs are in general following a "king of the hill" behaviour. They
will gang up against the biggest player, unless that player happens to
be their ally and he is not building a spaceship. The AI is quite
trusting towards its allies, and will not break alliances without
provokation (spaceship and not helping in wars being provokations).

AIs will help their allies if and only if they are in war with their
enemies. If you wish to give gifts to the AI in order to increase your
standing with that AI, open a diplomacy dialog and give only beneficial
clauses.