This patch provides a function distribute() in the utility code:
void distribute(int number, int groups, int *ratios, int *result);

This will distribute "number" elements into "groups" groups with ratios
given by the elements in "ratios". The resulting division is put into
the "result" array.

The only user (so far) is the trade disribution code. In this case
"number" is the amount of trade, "groups" is 3, and ratios[3] =
{sci_rate, tax_rate, lux_rate}. At the end result[] holds the amount of
sci/tax/luxury.

The algorithm used to determine the distribution is Hamilton's Method.

Patch by Jason in PR#9755