(PR#39427) BUG: city_style & unit_class name translation
Speaking of problems fixed in trunk, I forgot to include my fix of Qn_() 
on all of the rule names.  Not strictly needed with current data, this 
is an idea that might avoid problems in the future.  Everywhere Q_() is 
used to translate, Qn_() should be used to display untranslated data.