(PR#39437) attribute.c serialization error
Extensive error checking on saved and restored attribute "blocks" and "chunks".  
Still relies too much on asserts instead of actual remediation and recovery, 
but doesn't crash....

Fixed a number of initialization errors.  Yes, you have to clear the length, 
even when the data is already NULL.  They are separate (and this kind of error 
is likely the underlying cause of the crashes).

Also fixed another trivial enable-debug compilation error in packhand.c.

trunk revision 13068.