(PR#40096) deterministic borders (part 2)
This should expand city vision to the same limit as the borders, or
where the borders would be over the ocean.

Unlike the previous (part 1), this no longer yields exactly the same
borders.  The existing range progression (for borders 4) seemed a little
strange, and expanded much too quickly, so I've made it more uniform,
with every size having some effect.