API: Add Tile,Player:units_iterate(), and Player:cities_iterate()
Add iteration methods for Tile units and Player units and cities.

Iteration is safe w.r.t the linked list, because we copy it in lua
first.

We define two new API types, Unit_List_Link and City_List_Link,
both represent struct genlist_link (opaquely). The different types
are type-safe (can not be mixed) in the Lua API even if the underlying
type is the same.

It is safe to pass NULL to the genlist_link functions, so
SCRIPT_ASSERT is not needed.

See gna patch #1606