Don't drop a request for diplomat/spy actions based on client knowledge
When the player request that a unit do an action against the tile it is located
on key_unit_diplomat_actions() is called. It will check that the spy/diplomat
can do the action using the client's knowledge. If the unit can't it will drop
the request. If the unit can the request will be forwarded to a queue waiting to
ask the server. A rule may enable/disable an action based on data the player,
and therefore the client, can't see. The client could therefore drop a request
when there exist an action the unit legally can do. Since
key_unit_diplomat_actions() anyway will ask the server about what actions are
enabled the test can safely be removed.

See bug #22468