Bribe Unit: Diplomat battle any other capable unit on the target's tile.
When trying to bribe a target unit that shares a tile with any other unit
capable of defending in a diplomat battle the other unit will do a
diplomatic battle to stop it. The target unit of the bribe it self won't
defend in a diplomatic battle.

Requested by Marko Lindqvist <cazfi>

See patch #5300