Client: Make the sanity test in key_unit_trade_route() stricter.
The common client code's function key_unit_trade_route() checks that the actor
unit given to it by the client has the TradeRoute unit type flag before
forwarding the request to the server. This way the server won't be bothered by
theoretically impossible requests.

Because action enablers now control the establish trade route action it is
possible to have a unit type with the TradeRoute flag that never can establish a
trade route. Change the test to check for the unit type's ability to do the
establish trade route action instead.

See patch #5516