Have the orders system support server side target extra finding for pillage.
Pillage support was added to the orders system in patch #5861. When set to
FALSE the ruleset option pillage_select forbids the player to select what
target extra to pillage. Supporting pillage in the orders system without
server side target selection is therefore useless for clients that are
trying to support all rulesets.

See bug #23339