Add research_reqs, a requirement vector for research, to techs.
A tech's research_reqs must be true before it can be researched. For shared
research it is enough that it is true for one of the players sharing.

Only unchanging requirements are supported for now. This includes requiring
that a player must have a certain nationality to research a tech.

Tech requirements have their own warning telling the ruleset author to put
them in req1, req2 or root_req in stead.

See patch #5944