Fix unit state legality hostile actions cache.
Patch #5975 gave it a location and cleared it on initialization but didn't write
to it when a unit state permitted a hostile action. This made it look like
hostile actions always were forbidden in every unit state. Fix it by registering
what unit states allow a hostile action.

See bug #23464